agent 03/requirements_verification.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
requests